requirements.txt
asgiref==3.8.1
Django==4.2.13
django-filter==2.4.0
djangorestframework==3.15.1
sqlparse==0.5.0
tzdata==2024.1
django-filter==2.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1